requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.56.1
idna==3.4
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
